create middlewares module and use it in tests - fix some tests
