+++ src/actions/__tests__/users.js
@@ -1,17 +1,17 @@
 import configureMockStore from 'redux-mock-store';
-import thunk from 'redux-thunk';
 import axios from 'axios';
 import httpAdapter from 'axios/lib/adapters/http';
 import nock from 'nock';
 
 import { fetchUsers } from '../users';
+import { simpleActionMiddleware } from '../../helpers/storeMiddlewares';
 
 const host = 'http://localhost';
 
 axios.defaults.host = host;
 axios.defaults.adapter = httpAdapter;
 
-const mockStore = configureMockStore([thunk]);
+const mockStore = configureMockStore([simpleActionMiddleware]);
 
 describe('fetch users data', () => {
   const response = [{ id: '1', name: 'Welly' }];
@@ -32,7 +32,7 @@ describe('fetch users data', () => {
     ];
     const store = mockStore({ list: null });
 
-    store.dispatch(fetchUsers(`${host}/test`)).then(() => {
+    store.dispatch(fetchUsers()).then(() => {
       expect(store.getActions()).toEqual(expectedActions);
     });
   });
